requirements.txt
streamlit
requests
sqlalchemy
python-dotenv